zenoh_cognitive_framework/requirements.txt
eclipse-zenoh>=1.4.0
requests>=2.31.0
sympy>=1.12.0 